fix(ui): resolve bugs, remove tooltip auto-provider, compose refs
- zoom-image: render caption prop instead of ignoring it
- files: fix className merging in File component (cn item cva fix)
- tooltip: remove auto-TooltipProvider per shadcn pattern
- loading-skeleton: add local @keyframes shimmer to fix broken animation
- calendar: extract nested component definitions to module level
- sidebar: replace Math.random() with deterministic default width
- code-block: compose external ref with internal textInput ref
- tag-list: use cn() instead of template literal
- experience-entry: add TooltipProvider to tech stack list
