+++ packages/ui/src/index.ts
@@ -1,4 +1,1 @@
-import { cx } from 'class-variance-authority';
-import { twMerge } from 'tailwind-merge';
-
-export const cn = (...inputs: Parameters<typeof cx>) => twMerge(cx(inputs));
+export { cn } from './lib/cn';
